+++ app/src/main/java/com/bicycleApp/TourDetailsActivity.java
@@ -1,10 +1,13 @@
 package com.bicycleApp;
 
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
+import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.MenuItem;
 import android.view.View;
@@ -14,6 +17,7 @@ import android.widget.Toast;
 import com.google.android.material.appbar.MaterialToolbar;
 
 import Data.MyDatabase;
+import Utils.Utilities;
 
 public class TourDetailsActivity extends AppCompatActivity {
 
@@ -22,6 +26,7 @@ public class TourDetailsActivity extends AppCompatActivity {
     private double time, distance;
     private int id;
     private TextView titleText, timeText, distanceText, speedText;
+    private String title;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -36,12 +41,13 @@ public class TourDetailsActivity extends AppCompatActivity {
         timeText = findViewById(R.id.text_tour_details_time);
         distanceText = findViewById(R.id.text_tour_details_distance);
         speedText = findViewById(R.id.text_tour_details_avg_speed);
-        titleText.setText(this.getIntent().getStringExtra("Title"));
+        title = getIntent().getStringExtra("Title");
+        titleText.setText(title);
         int hours = (int) time % 86400 / 3600;
         int minutes = (int) time % 86400 % 3600 / 60;
         int seconds = (int) time % 86400 % 3600 % 60;
         timeText.setText(String.format("%02d:%02d:%02d", hours, minutes, seconds));
-        distanceText.setText("" + distance +" km");
+        distanceText.setText("" + Utilities.roundTo2DecimalPlace(distance) +" km");
         speedText.setText(""+Utilities.roundTo2DecimalPlace(distance/(time/3600))+" km/h");
 
         toolbar.setTitle(this.getIntent().getStringExtra("Date").substring(0,this.getIntent().getStringExtra("Date").length()-9));
@@ -55,14 +61,47 @@ public class TourDetailsActivity extends AppCompatActivity {
         toolbar.setOnMenuItemClickListener(new Toolbar.OnMenuItemClickListener() {
             @Override
             public boolean onMenuItemClick(MenuItem item) {
-                try {
-                    deleteItem();
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
+                if(item.getTitle().equals("delete")){
+                    try {
+                        deleteItem();
+                    } catch (InterruptedException e) {
+                        e.printStackTrace();
+                    }
                 }
+                if(item.getTitle().equals("route"))
+                    showRoute();
+                if(item.getTitle().equals("edit"))
+                    editItem();
+
                 return false;
             }
         });
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        if (requestCode == 1 && resultCode == Activity.RESULT_OK)
+            titleText.setText(data.getStringExtra("TitleResult"));
+        super.onActivityResult(requestCode, resultCode, data);
+    }
+
+    private void showRoute(){
+        Intent intent = new Intent(this, MapShowRouteActivity.class);
+        intent.putExtra("TourId", id);
+        intent.putExtra("StartLat", getIntent().getDoubleExtra("StartLat",0));
+        intent.putExtra("StartLon", getIntent().getDoubleExtra("StartLon",0));
+        intent.putExtra("EndLat", getIntent().getDoubleExtra("EndLat",0));
+        intent.putExtra("EndLon", getIntent().getDoubleExtra("EndLon",0));
+        startActivity(intent);
+    }
+
+    private void editItem(){
+        Intent intent = new Intent(this, TourEditActivity.class);
+        intent.putExtra("TourId", id);
+        intent.putExtra("Time", time);
+        intent.putExtra("Distance", distance);
+        intent.putExtra("Title", title);
+        startActivityForResult(intent, 1);
     }
 
     private void deleteItem() throws InterruptedException {
